Moved Management API classes into ASCOM.Common
