+++ ASCOM.Common/ASCOM.Common.Alpaca/Discovery/Entity Classes/AlpacaConfiguredDevice.cs
