+++ ASCOM.Common/ASCOM.Common.Alpaca/Discovery/Entity Classes/AlpacaDeviceDescription.cs
